% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_name.R
\name{date_name}
\alias{date_name}
\title{Format a date into a consistent human-readable format}
\usage{
date_name(x, abbr_day = TRUE, abbr_month = TRUE, include_year = FALSE)
}
\arguments{
\item{x}{a date in date format.}

\item{abbr_day}{logical. Whether or not you would like the name of the day
of the week to be abbreviated. Defaults to TRUE.}

\item{abbr_month}{logical. Whether or not you would like the name of the month
to be abbreviated. Defaults to TRUE.}

\item{include_year}{logical. Whether you would like the date to include the year.
Defaults to FALSE.}
}
\value{
Returns a brief commentary (including the figure for
increases or decreases) as a string
}
\description{
Format a date into a consistent human-readable format
}
