% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_change.R
\name{number_change}
\alias{number_change}
\title{Return a commentary snippet for one or more values.
Formats the numbers as a comma separated numeric and provides
simple up/down commentary}
\usage{
number_change(x, description = "up", unchanged_limit = 100, ...)
}
\arguments{
\item{x}{a vector of numeric values.}

\item{description}{code of descriptive words used in commentary.
Defaults to "up"; up/down/unchanged.}

\item{unchanged_limit}{numeric, the value below which you consider
the parameter to not represent a change in either direction. Defaults to 100.}

\item{...}{additional arguments to pass to the scales::comma function}
}
\value{
Returns a brief commentary (including the figure
for increases or decreases) as a string
}
\description{
Return a commentary snippet for one or more values.
Formats the numbers as a comma separated numeric and provides
simple up/down commentary
}
