% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round}
\alias{round}
\title{Round numbers according to statistical principles rather than data science ones
i.e. numbers ending in a 5 will always be rounded up, regardless of preceeding number}
\usage{
round(x, digits = 0)
}
\arguments{
\item{x}{a numeric value}

\item{digits}{integer indicating the number of decimal places.
Negative values are allowed; Rounding to a negative number of digits means rounding to a power of ten, so for example round(x, digits = -2) rounds to the nearest hundred.}
}
\description{
Round numbers according to statistical principles rather than data science ones
i.e. numbers ending in a 5 will always be rounded up, regardless of preceeding number
}
