% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_descriptions}
\alias{show_descriptions}
\alias{show_format}
\title{Return a data table of the description codes and examples of their use}
\usage{
show_format()
}
\description{
Return a data table of the description codes and examples of their use
}
