% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_number.R
\name{single_number}
\alias{single_number}
\title{Return a commentary snippet for any number value.
Formats the number with comma separation
and provides simple up/down commentary}
\usage{
single_number(x, description = "up", unchanged_limit = 100, ...)
}
\arguments{
\item{x}{a numeric value.}

\item{description}{code of descriptive words used in commentary.
Defaults to "up"; up/down/unchanged.}

\item{unchanged_limit}{numeric, the value below which you
consider the parameter to not represent a change in either direction.
Defaults to 100.}

\item{...}{additional arguments to pass to the scales::comma function}
}
\value{
Returns a brief commentary (including the figure for
increases or decreases) as a string
}
\description{
Return a commentary snippet for any number value.
Formats the number with comma separation
and provides simple up/down commentary
}
