% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smart_paste}
\alias{smart_paste}
\title{Paste together objects based on how many of them there are
Uses commas between everything except the last value}
\usage{
smart_paste(x)
}
\arguments{
\item{x}{a vector of two or more objects}
}
\description{
Paste together objects based on how many of them there are
Uses commas between everything except the last value
}
